refactor(packs): add service layer with public business functions
- Created service.go with 3 public functions:
  - FindPackByID: wrapper for repository function (used by images package)
  - CheckPackOwnership: wrapper for repository function (used by images package)
  - FindPackIDByPackName: utility function for pack lookup
- Made findPackByID and checkPackOwnership private in repository.go
- Removed FindPackIDByPackName from packs.go (moved to service.go)
- Follows DDD pattern: service layer (public) wraps repository layer (private)
- All tests pass (46.1% coverage maintained)
